Added priority test

+++ Userland/UserModule/standard_library/include/mystdlib.h
@@ -13,7 +13,7 @@
 
 size_t numToStr(long value, char* target, size_t base);
 void numToStrSized(long value, char* target, size_t base, size_t length);
-long strToNum(char* string);
+long long strToNum(char* string);
 long strToNumPos(char* string);
 int strcmp(char* s1, char* s2);
 size_t strlen(char *string);
